requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
Django==3.2.7
django-environ==0.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
passlib==1.7.4
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
